Created a protected instance of WebDriverWait in WebPageHelpers, and removed all of the declarations and assignments from all of the other pageObjects. Also removed repeat assignments to the wait object.
Added a 'setName' method to Community, this takes the community name as
it appears in the url and assigns it to a String 'name'.

Added some more CSSSelectors to createAccount.

Added a method to WebPageHelpers waitForPageToLoad which takes a
cssSelector and waits until the element appears. Useful for using
directly in the implementations of the 'Given', 'When', 'Then' blocks.

Added steps to the user story T121.